wow pending list iz working 1/2

+++ Und_System/Data/CGLCreditDebit.cs
@@ -11,10 +11,10 @@ namespace Und_System.Data
         public required string DCHS_To { get; set; }
         public DateTime Period_From { get; set; }
         public DateTime Period_To { get; set; }
-        public float Gross_Pay { get; set; }
-        public float Net_Pay { get; set; }
+        public double Gross_Pay { get; set; }
+        public double Net_Pay { get; set; }
         public int Total_Application { get; set; }
-        public float Total_Coverage { get; set; }
+        public double Total_Coverage { get; set; }
         public DateTime Submitted_Date { get; set; }
         public required string Submitted_By { get; set; }
         public required string Receipt { get; set; }
